fix warping on scroll
we were moving the backbuffer by a whole number of bins, but the FFT
params don't necessarily change by a multiple of the bin width. When
scrolling with the VFO knob that error would accumulate over lots of
scroll events and lead to misalignment.

Trying to scroll by fractional bins using linear interpolation turns the
history into a smudgy mess, but we can accumulate the error and make
sure that we're always within one bin of correct, which looks fine.